add prompt and configuration key
