+++ package.nls.json
@@ -25,5 +25,6 @@
     "azureApiManagement.showProductPolicy": "Edit Product policy",
     "azureApiManagement.showArmProduct": "Edit Product",
     "azureApiManagement.openWorkingFolder": "Open Working Folder",
-    "azureApiManagement.setupWorkingFolder": "Setup Working Folder"
+    "azureApiManagement.setupWorkingFolder": "Setup Working Folder",
+    "azureApiManagement.showOpenWorkingFolderPrompt": "Show prompt to open or initialize working folder during policy editing."
 }
