+++ src/commands/setupWorkingFolder.ts
@@ -7,6 +7,7 @@ import * as fse from 'fs-extra';
 import * as path from 'path';
 import { IActionContext } from 'vscode-azureextensionui';
 import { ext } from '../extensionVariables';
+import { localize } from '../localize';
 import { cpUtils } from '../utils/cpUtils';
 import { dotnetUtils } from '../utils/dotnetUtils';
 import { getDefaultWorkspacePath } from '../utils/fsUtil';
@@ -17,8 +18,10 @@ export async function setupWorkingFolder(this: IActionContext): Promise<void> {
    // cleanup the directory, incase it was previously setup or used.
    await fse.emptyDir(workingFolderPath);
 
+   // make sure dotnet tools are installed.
    await dotnetUtils.validateDotnetInstalled(this);
 
+   // run dotnet new webapp to create a dummy empty asp.net webapp project to work with razor files.
    await cpUtils.executeCommand(
        ext.outputChannel,
        workingFolderPath,
@@ -29,9 +32,13 @@ export async function setupWorkingFolder(this: IActionContext): Promise<void> {
    // Copy the supporting files.
    await fse.copy(ext.context.asAbsolutePath(path.join('resources', 'projectFiles')), workingFolderPath, { overwrite: true, recursive: false });
 
+   // run dotnet build atleast once to get the intellisense working in razor files.
    await cpUtils.executeCommand(
     ext.outputChannel,
     workingFolderPath,
     'dotnet',
     'build');
+
+   ext.outputChannel.appendLine(localize("folderInitialized", "Working folder initialization completed.", workingFolderPath));
+   ext.outputChannel.show();
 }
